backend/05_INTAKE/requirements.txt
supabase
fastapi
uvicorn
python-multipart
python-dotenv
pydantic
google-generativeai
psycopg2-binary
